CR - Organize NavigationMenu components under own folder
